Prevent the application to take all CPU usage.

+++ preprocess/pdf_preprocess.py
@@ -1,30 +1,32 @@
-import cv2
 import os
 import pytesseract
+import psutil
 from concurrent.futures import ThreadPoolExecutor
-
 from pdf2image import convert_from_path
 from pypdf import PdfMerger
-
-from image_preprocess import noise_removal, downscale_image
-
 from PIL import Image
 
 Image.MAX_IMAGE_PIXELS = None
+
+# Function to get CPU usage percentage
+
+# Set the maximum allowed CPU usage percentage
+max_cpu_usage = 90  # Adjust this value as needed
+
+
+def get_cpu_usage():
+    return psutil.cpu_percent(interval=1)
+
+
+width, height = 800, 1200
 
 
 def preprocess_page(page_num, image, processed_dir_path, extracted_dir_path):
     try:
         print(f"--- Processing page {page_num + 1}... ---")
 
-        # image = cv2.imread(unprocessed_dir_path)
-
-        # enhanced_img = noise_removal(unprocessed_dir_path)
-
-        # Save the processed image
-        # processed_image_path = os.path.join(
-        #     processed_dir_path, f"page_{page_num + 1}.png")
-        # cv2.imwrite(processed_image_path, enhanced_img)
+        # Resize image to an optimal size
+        image = image.resize((width, height), Image.ANTIALIAS)
 
         # Perform OCR on the processed image
         extracted_text = pytesseract.image_to_string(
@@ -39,7 +41,6 @@ def preprocess_page(page_num, image, processed_dir_path, extracted_dir_path):
                 extracted_dir_path, f"page_{page_num + 1}.txt")
             with open(extracted_text_path, 'w', encoding='utf-8') as text_file:
                 text_file.write(extracted_text)
-                text_file.close()
             print("Done saving extracted text!")
 
         # convert to searchable PDF
@@ -74,15 +75,25 @@ def preprocess_pdf(pdf_path, processed_pdf_dir, extracted_text_dir, searchable_p
 
     pages = convert_from_path(pdf_path, thread_count=4, dpi=150, grayscale=True, hide_annotations=True)
     print("Done converting to images!")
+
     # PDF merger
     merger = PdfMerger()
 
-    # Set up a thread pool with a specified number of workers
-    with ThreadPoolExecutor(max_workers=4) as executor:
+    # Set up a thread pool with an optimal number of workers
+    max_workers = psutil.cpu_count(logical=False) or 1  # Use physical cores if available, otherwise use 1
+    processing_successful = True  # Flag to indicate if processing was successful for all pages
+    with ThreadPoolExecutor(max_workers=max_workers) as executor:
         try:
             # Process each page using thread pool
             pdf_page_paths = []
             for page_num in range(len(pages)):
+                # Check CPU usage before processing each page
+                current_cpu_usage = get_cpu_usage()
+                if current_cpu_usage >= max_cpu_usage:
+                    print(f"CPU usage reached {current_cpu_usage}%, stopping processing.")
+                    processing_successful = False
+                    break
+
                 image = pages[page_num]
 
                 # save the image
@@ -94,7 +105,7 @@ def preprocess_pdf(pdf_path, processed_pdf_dir, extracted_text_dir, searchable_p
                 total_pixels = image.size[0] * image.size[1]
 
                 if total_pixels > 178956970:
-                    image = downscale_image(image)
+                    image = image.resize((width, height), Image.ANTIALIAS)
 
                 future = executor.submit(preprocess_page, page_num, image, processed_dir_path_after,
                                          extracted_dir_path)
@@ -103,17 +114,22 @@ def preprocess_pdf(pdf_path, processed_pdf_dir, extracted_text_dir, searchable_p
             # Wait for all PDF pages to be generated
             pdf_page_paths = [future.result() for future in pdf_page_paths if future.result() is not None]
 
-            # PDF merger
-            merger = PdfMerger()
-            for pdf_page_path in pdf_page_paths:
-                # Append the PDF page to the merger
-                merger.append(pdf_page_path)
-                os.remove(pdf_page_path)
+            # Continue with PDF merging even if CPU usage threshold was exceeded
+            if processing_successful:
+                # PDF merger
+                merger = PdfMerger()
+                for pdf_page_path in pdf_page_paths:
+                    # Append the PDF page to the merger
+                    merger.append(pdf_page_path)
+                    os.remove(pdf_page_path)
 
-            # convert the processed images to searchable PDF
-            searchable_pdf_dir_with_name = os.path.join(searchable_pdf_dir, file_name + ".pdf").replace("\\", "/")
+                # convert the processed images to searchable PDF
+                searchable_pdf_dir_with_name = os.path.join(searchable_pdf_dir, file_name + ".pdf").replace("\\", "/")
 
-            merger.write(searchable_pdf_dir_with_name)
-            merger.close()
+                merger.write(searchable_pdf_dir_with_name)
+                merger.close()
+                print("Searchable PDF created successfully.")
+            else:
+                print("Processing interrupted due to high CPU usage. Searchable PDF not created.")
         except Exception as e:
             print("Error in preprocess_pdf: ", e)
